requirements.txt
# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# LLM & Embeddings
openai==1.65.2
sentence-transformers==3.4.1

# Vector Store
faiss-cpu==1.9.0.post1

# Document Parsing
pdfplumber==0.11.5
openpyxl==3.1.5

# Data & Storage
sqlalchemy==2.0.38
pandas==2.2.3
numpy==2.2.3

# Utilities
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.0
httpx==0.28.1
python-multipart==0.0.20
aiofiles==24.1.0

# Sample data generation
fpdf2==2.8.2
